Refs #920: Proof of concept https
This is a proof of concept to show that we could use https.  I changed
all the instances of http to https, added a dummy self-signed
certificate, and it worked pretty much out of the box.  You'll obviously
have to accept the usual security warning in your browser about
self-signed certificates.

I wasn't sure if the websockets connections were encrypted too, or if
they were in a side channel for which encryption needs to be configured
too.  From Wireshark, all I could see while typing in the terminal was
traffic towards port 3000, so I assume they are tunneled in the https
connection.

This is just a proof of concept, so we can discuss what to do next.